Updated readme with references, changed main wokring file to "newton.py"
